// src/types/Spells.ts
export interface SpellCardProps {
  school: string;
  title: string;
  description: string;
  nivel: string;
  timeConjuration: string;
  alcance: string;
  componentes: string;
  duration: string;
  ritual: boolean;
  bard: boolean;
  wizard: boolean;
  cleric: boolean;
  druid: boolean;
  sorcerer: boolean;
  mage: boolean;
  paladin: boolean;
  ranger: boolean;
}